update: 1.0.1
https://www.openwall.com/lists/announce/2019/06/30/2

Date: Sun, 30 Jun 2019 18:01:23 +0200
From: Solar Designer <solar@...nwall.com>
To: announce@...ts.openwall.com
Subject: [openwall-announce] yespower 1.0.1

Hi,

yespower is a proof-of-work (PoW) focused fork of yescrypt.  While
yescrypt is a password-based key derivation function (KDF) and password
hashing scheme, and thus is meant for processing passwords, yespower is
meant for processing trial inputs such as block headers (including
nonces) in PoW-based blockchains.

This is to announce a minor update of yespower released today as 1.0.1,
and available on the project's homepage:

https://www.openwall.com/yespower/

The changes are described in the included change log as follows:

---
	Changes made between 1.0.0 (2018/07/12) and 1.0.1 (2019/06/30).

Fill the destination buffer with all set bits on error for fail-safety
of the caller's "< target" check in case the caller neglects to check
for errors.

Simplified SMix2 for its final invocation with Nloop=2 in yespower 0.5.

Revised the "XOR of yespower" tests to trigger duplicate index in the
last SMix2 invocation in yespower 0.5 for N=2048 with at least one of
the values of r being tested.  This is needed to test that a proper
kind of BlockMix is used in that special case, which would previously be
left untested.

Added x32 ABI support (x86-64 with 32-bit pointers).

Added a bit more detail to the README on caching of the computed PoW
hashes when integrating yespower in an altcoin based on Bitcoin Core.
---

Alexander